Prepare for 0.5.0 release (#19)
Updates to 0.5.0 and include bug fixes associated with the merger of local mode and feedback code

Included bug fixes:
- Updates API endpoints
    - Adds `is_openai` to `/service` return
    - Removes 501 returns for disclaimer endpoints and instead returns "successful" values
    - Adds `is_openai` case to feedback endpoint that returns an "unsupported" message
- Adds a helper function to wipe `lastPrompt` value when the value becomes irrelevant
- Updates model dropdown list to remember the current model and wipes the lastPrompt value on model update
- Saves the `is_openai` value when the service url is set and only shows the feedback UI when `is_openai` is false

Weird UI edge case:
The feedback buttons are not immediately removed when the service UI is switch to an OpenAI one. The status bar is updated when you change tabs and the notebook cell toolbar is updated either when switching cells or running a completion. This is not an issue though since if the user tries to use the feedback during this edge case they will see a "unsupported" notification.